+++ resources/assets/js/components/enso/select/VueSelect.vue
@@ -79,7 +79,7 @@
 
 <script>
 
-import { debounce } from 'lodash';
+import debounce from 'lodash/debounce';
 import { mapGetters } from 'vuex';
 import vClickOutside from 'v-click-outside';
 import fontawesome from '@fortawesome/fontawesome';