+++ resources/assets/js/components/enso/vuedatatable/responsive/ResponsiveTable.js
@@ -1,16 +1,14 @@
+import debounce from 'lodash/debounce';
+
 class ResponsiveTable {
     constructor(el, context) {
         this.el = el;
         this.context = context;
-        this.hiding = false;
+        this.resize = debounce(this.resize, 16);
         this.width = null;
     }
 
-    updateSize() {
-        this.width = this.el.offsetWidth;
-    }
-
-    shouldResize() {
+    hasChanged() {
         return this.width !== this.el.offsetWidth;
     }
 
@@ -19,10 +17,14 @@ class ResponsiveTable {
     }
 
     shouldShow() {
-        return this.el.offsetWidth === this.el.scrollWidth;
+        return this.el.offsetWidth === this.el.scrollWidth && this.hasChanged();
     }
 
-    hideColumns() {
+    updateWidth() {
+        this.width = this.el.offsetWidth;
+    }
+
+    hideColumn() {
         const column = this.context.template.columns
             .filter(column => column.meta.visible && !column.meta.hidden && !column.meta.rogue)
             .pop();
@@ -31,9 +33,9 @@ class ResponsiveTable {
             return;
         }
 
-        this.hiding = true;
         column.meta.hidden = true;
-        this.retryFit();
+
+        this.update();
     }
 
     showColumn() {
@@ -45,32 +47,21 @@ class ResponsiveTable {
         }
 
         column.meta.hidden = false;
-        this.retryFit();
+        this.update();
     }
 
-    resize() {
+    update() {
         this.context.$nextTick(() => {
-            if (this.shouldResize()) {
-                this.fit();
+            this.updateWidth();
+            if (this.shouldHide()) {
+                this.hideColumn();
             }
         });
     }
 
-    retryFit() {
-        this.context.$nextTick(() => {
-            this.updateSize();
-            this.fit();
-        });
-    }
-
-    fit() {
+    resize() {
         if (this.shouldHide()) {
-            this.hideColumns();
-            return;
-        }
-
-        if (this.hiding) {
-            this.hiding = false;
+            this.hideColumn();
             return;
         }