+++ resources/assets/js/components/enso/vuedatatable/responsive/vResponsive.js
@@ -1,19 +1,11 @@
-import resizeDetector from 'element-resize-detector';
+import ResizeSensor from 'css-element-queries/src/ResizeSensor';
 import ResponsiveTable from './ResponsiveTable';
 
-const erd = resizeDetector({ strategy: 'scroll' });
 let table = null;
 
 export default {
     inserted: (el, binding, { context }) => {
         table = new ResponsiveTable(el, context);
-        erd.listenTo(el, () => table.resize());
-    },
-    componentUpdated: () => {
-        table.resize();
-    },
-    unbind(el) {
-        const erd = resizeDetector({ strategy: 'scroll' });
-        erd.uninstall(el);
+        const sensor = new ResizeSensor(el, () => table.resize());
     },
 };